feat: code is divided into modules
